fix: rename file and add missing test
